% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_auth_header}
\alias{add_auth_header}
\title{Add header to request}
\usage{
add_auth_header(req)
}
\arguments{
\item{req}{An http request, derived from \code{httr2::request}.}
}
\value{
A modified request, with the authentication token embedded in the request header.
}
\description{
The authorization seeks an environment variable called \code{AUTOCODE_API_KEY} for
the authentication token.
}
