% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_service_record_filter}
\alias{adjust_service_record_filter}
\title{Adjust the service record filter}
\usage{
adjust_service_record_filter(filter)
}
\arguments{
\item{filter}{Define the filter to subset the results. A single character
value. This is one of '', 'pvp', 'ranked', 'bots', or 'custom', for
convenience, which is modified by the function to match the API spec.}
}
\value{
A modified character value.
}
\description{
The service record API has a few filters that can be passed:
\itemize{
\item None (all)
\item matchmade:pvp
\item matchmade:social
\item matchmade:ranked
\item matchmade:bots
\item custom
}

For convenience, the user inputs one of \verb{'all', 'pvp', 'social', 'ranked', 'bots', 'custom'}
}
