% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{api_base_url}
\alias{api_base_url}
\title{API URL}
\usage{
api_base_url()
}
\value{
A character vector with the base API url.
}
\description{
The base API url. The default version is currently
0.3.9. This version is retrieved via an R
option named \code{halodotapi.default_version}
}
