% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{api_player_csr}
\alias{api_player_csr}
\title{CSR API URL}
\usage{
api_player_csr(
  gamertag,
  season,
  input = c("mnk", "controller", "crossplay"),
  queue
)
}
\arguments{
\item{gamertag}{Character value for the gamertag to view.}

\item{season}{Numeric value for the season to view.}

\item{input}{Character value, one of 'mnk', 'controller', or 'crossplay' for the specific input type.}

\item{queue}{Character value, one of 'open', or 'solo-duo' for the queuing method.}
}
\description{
Returns the formatted API URL for the player's CSR data.
}
\details{
Note that solo-duo queue only has input values for 'mnk' and 'controller',
whereas open queue only has the 'crossplay' input value.
}
