% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{api_player_list_matches}
\alias{api_player_list_matches}
\title{Match Record API URL}
\usage{
api_player_list_matches(gamertag, count = NULL, offset = NULL, mode = NULL)
}
\arguments{
\item{gamertag}{Character value for the gamertag to view.}

\item{page}{Numeric value for the page to view. The data returned is paginated,
and you will want to use the next page number that is returned in the response.}
}
\description{
Returns the formatted API URL for the player's last 100 matches.
}
