% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{api_player_service_record}
\alias{api_player_service_record}
\title{Service Record API URL}
\usage{
api_player_service_record(gamertag)
}
\arguments{
\item{gamertag}{Character value for the gamertag to view.}
}
\description{
Returns the formatted API URL for the player's service record.
}
