% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{get_match_details}
\alias{get_match_details}
\title{Get match session details}
\usage{
get_match_details(match.id)
}
\arguments{
\item{match.id}{The 32-character match identifier, typically retrieved from a player's match history.}
}
\value{
A list of three data frames, containing match session information, player stats, and player medals
}
\description{
This query takes no gamertag input, instead retrieving some high-level
information about the match (1 record for several match-level variables), and
all players in the match (1 record for each player).
}
\details{
There are three items returned:
\itemize{
\item Match session details
\item Player details
\item Player medals
}

The match session details contain information on dimensions related to that
specific game session. These include:
\itemize{
\item Match ID
\item Play date-time
\item Game mode
\item Map name
}

And other dimensions.

The player details contain individual performance summaries for the match,
for each participant in the match. These are many of the same dimensions as
provided in \code{get_player_match_history} and \code{get_player_service_record}. This
players dataset has a many-to-one relationship with the match ID.

The player medals dataset has a many-to-one relationship with the player
details, and a many-to-one relationship with the match ID. This dataset includes:
\itemize{
\item Medal-specific identifier codes
\item Medal names
\item Frequency of medals earned
\item Match ID and gamertag of the recipient.
}

Most datasets include thumbnails of the assets involved.
}
