% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{get_player_csr}
\alias{get_player_csr}
\title{Get CSR data}
\usage{
get_player_csr(gamertag, season, handle)
}
\arguments{
\item{gamertag}{Character value for the gamertag to view.}

\item{season}{Numeric value for the season to view.}

\item{handle}{\code{curl} handle to use, created with curl::new_handle. Used to supply authentication token.}
}
\description{
Automatically loops through valid input and queue parameters, aggregating the
results together in a data frame.
}
