% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{get_player_service_record}
\alias{get_player_service_record}
\title{Get player service record}
\usage{
get_player_service_record(
  gamertag,
  type = c("multiplayer", "campaign"),
  filter = c("all", "pvp", "social", "ranked", "bots", "custom")
)
}
\arguments{
\item{gamertag}{Gamertag to query. A single character string.}

\item{type}{One of either \code{multiplayer} or \code{campaign}.}

\item{filter}{One of \verb{'all', 'pvp', 'social', 'ranked', 'bots', 'custom'}.}
}
\value{

}
\description{
The player service record is a record of medals, and single overall
statistics over the player's lifetime.
}
\details{
Examples of these are lifetime totals of:
\itemize{
\item Kills
\item Deaths
\item Assists
\item Betrayals
\item Hijacks
\item Damage dealt/taken
\item Shots fired/missed/accuracy
\item KDA/KDR
\item Matches played
\item Time played
\item Score
}

and several more dimensions.
}
