% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_NULL_with_NA}
\alias{replace_NULL_with_NA}
\title{Replace NULLS in a nested list with NA}
\usage{
replace_NULL_with_NA(x, depth)
}
\arguments{
\item{x}{A nested list, possibly with null values}
}
\value{
The same nested list, but with NULL's substituted for NA's
}
\description{
Converting a nested list to a data frame requires that NULL's be converted to
NA values. This is because the implied number of rows for that entry is 0,
leading to a mismatch in the number of rows among all entries in the dataset.
To resolve, replace the NULL's with NA, which have length.
}
