% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{get_player_match_data}
\alias{get_player_match_data}
\title{Get match data}
\usage{
get_player_match_data(gamertag, handle, one = FALSE)
}
\arguments{
\item{gamertag}{Character value for the gamertag to view.}

\item{handle}{\code{curl} handle to use, created with curl::new_handle. Used to supply authentication token.}

\item{one}{Boolean, if TRUE returns only the first page. Useful if}
}
\description{
Runs through the pagination for the match history for the supplied gamertag.
Each response is returned as JSON, which is formatted as a data.frame for the
user. The response contains a field for the next page number, which is used
until it returns \code{NULL}.
}
